+++ depthai_nodes/node/gather_data.py
